feat: enable eslint language server support
This change will display diagnostics for each specified language.

The updated languages all support typescript-language-server by default,
but this needs to be specified to avoid overwriting the default language
servers array with the change.

This works because eslint settings are defined in Helix by default.

Note that due to a change in vscode-langservers-extracted, this needs to
use version 4.8.0. Homebrew has the latest, so I downgraded and pinned.