refactor: install vscode langservers via npm
Switching to use NPM to install the HTML/CSS/JS/TS language servers.

Installing via node seems to be the preferred installation method, plus
it makes it easier to specify the version, as the current release does
not work for ESLint, and a potential Helix fix has not yet been merged.

Additionally, pinning in Homebrew is a manual process so the Node method
is less error prone.

See here for curent state: https://github.com/helix-editor/helix/wiki/Language-Server-Configurations#eslint